S203: composition elevation sprint + deep debt execution
Standardize compute.dispatch wire contract — all 8 dispatch handlers share
canonical {domain, operation, job_id, status, output, error, metadata} envelope
for primalSpring typed extractors. Smart refactor 4 large files (test extraction).
Deprecate get_primal_default_port, evolve GPU buffer unsafe to NonNull pattern,
centralize discovery ports, resolve clippy suppressions, clean deny.toml advisories.
Document wire contract in specs/DISPATCH_WIRE_CONTRACT.md. Update all root docs to S203.

Made-with: Cursor

+++ crates/core/toadstool/src/biomeos_integration/auth/mod.rs
@@ -238,14 +238,13 @@ impl AuthenticationManager {
         if !self.config.signature_validation {
             return Ok("signature_disabled".to_string());
         }
-        #[allow(clippy::cast_sign_loss, clippy::cast_possible_wrap)]
         let payload = format!(
             "{}:{}:{}",
             token.id,
             target_primal,
             SystemTime::now()
                 .duration_since(SystemTime::UNIX_EPOCH)
-                .map(|d| d.as_secs() as i64)
+                .map(|d| d.as_secs())
                 .unwrap_or(0)
         );
         self.sign_payload(&payload).await
@@ -260,13 +259,12 @@ impl AuthenticationManager {
         if !self.config.signature_validation {
             return Ok("signature_disabled".to_string());
         }
-        #[allow(clippy::cast_sign_loss, clippy::cast_possible_wrap)]
         let payload = format!(
             "verify:{}:{}",
             primal_name,
             SystemTime::now()
                 .duration_since(SystemTime::UNIX_EPOCH)
-                .map(|d| d.as_secs() as i64)
+                .map(|d| d.as_secs())
                 .unwrap_or(0)
         );
         self.sign_payload(&payload).await
